refactor common test methods
