+++ pkg/bootstrap/steps/admin_password_test.go
@@ -126,10 +126,7 @@ func TestCreateCredentials(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			config, err := makeTestConfig(t, Config{}, tt.secret)
-			if err != nil {
-				t.Fatalf("error creating config: %v", err)
-			}
+			config := makeTestConfig(t, Config{}, tt.secret)
 
 			out, err := createCredentials(tt.input, &config)
 			if err != nil {
