+++ pkg/bootstrap/steps/common_tests.go
@@ -3,17 +3,14 @@ package steps
 import (
 	"testing"
 
-	"github.com/weaveworks/weave-gitops-enterprise/pkg/bootstrap/utils"
+	"github.com/weaveworks/weave-gitops-enterprise/test/utils"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
-func makeTestConfig(t *testing.T, config Config, objects ...runtime.Object) (Config, error) {
-	fakeClient, err := utils.CreateFakeClient(t, objects...)
-	if err != nil {
-		return Config{}, err
-	}
+func makeTestConfig(t *testing.T, config Config, objects ...runtime.Object) Config {
+	fakeClient := utils.CreateFakeClient(t, objects...)
 	cliLogger := utils.CreateLogger()
-	config = Config{
+	return Config{
 		KubernetesClient:   fakeClient,
 		Logger:             cliLogger,
 		WGEVersion:         config.WGEVersion,
@@ -24,5 +21,4 @@ func makeTestConfig(t *testing.T, config Config, objects ...runtime.Object) (Con
 		PrivateKeyPath:     config.PrivateKeyPath,
 		PrivateKeyPassword: config.PrivateKeyPassword,
 	}
-	return config, err
 }
