+++ pkg/bootstrap/utils/k8s_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/alecthomas/assert"
 	"github.com/loft-sh/vcluster/pkg/util/random"
+	"github.com/weaveworks/weave-gitops-enterprise/test/utils"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -18,7 +19,7 @@ func TestGetSecret(t *testing.T) {
 	secretNamespace := "flux-system"
 	invalidSecretName := "invalid-secret"
 
-	fakeClient, err := CreateFakeClient(t, &v1.Secret{
+	fakeClient := utils.CreateFakeClient(t, &v1.Secret{
 		ObjectMeta: metav1.ObjectMeta{Name: secretName, Namespace: secretNamespace},
 		Type:       "Opaque",
 		Data: map[string][]byte{
@@ -26,9 +27,6 @@ func TestGetSecret(t *testing.T) {
 			"password": []byte("test-password"),
 		},
 	})
-	if err != nil {
-		t.Fatalf("error creating fake client: %v", err)
-	}
 
 	secret, err := GetSecret(fakeClient, invalidSecretName, secretNamespace)
 	assert.Error(t, err, "error fetching secret: %v", err)
@@ -53,12 +51,9 @@ func TestCreateSecret(t *testing.T) {
 		"password": []byte("test-password"),
 	}
 
-	fakeClient, err := CreateFakeClient(t)
-	if err != nil {
-		t.Fatalf("error creating fake client: %v", err)
-	}
+	fakeClient := utils.CreateFakeClient(t)
 
-	err = CreateSecret(fakeClient, secretName, secretNamespace, secretData)
+	err := CreateSecret(fakeClient, secretName, secretNamespace, secretData)
 	assert.NoError(t, err, "error creating secret: %v", err)
 
 	secret, err := GetSecret(fakeClient, secretName, secretNamespace)
@@ -80,12 +75,8 @@ func TestDeleteSecret(t *testing.T) {
 		"password": []byte("test-password"),
 	}
 
-	fakeClient, err := CreateFakeClient(t)
-	if err != nil {
-		t.Fatalf("error creating fake client: %v", err)
-	}
-
-	err = CreateSecret(fakeClient, secretName, secretNamespace, secretData)
+	fakeClient := utils.CreateFakeClient(t)
+	err := CreateSecret(fakeClient, secretName, secretNamespace, secretData)
 	assert.NoError(t, err, "error creating secret: %v", err)
 
 	err = DeleteSecret(fakeClient, secretName, secretNamespace)
