+++ test/utils/utils.go
@@ -1,24 +1,31 @@
 package utils
 
 import (
+	"os"
 	"testing"
 
-	corev1 "k8s.io/api/core/v1"
+	"github.com/weaveworks/weave-gitops/pkg/kube"
+	"github.com/weaveworks/weave-gitops/pkg/logger"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
+// CreateFakeClient create fake client for testing with default wge schemes
 func CreateFakeClient(t *testing.T, clusterState ...runtime.Object) client.Client {
 	t.Helper()
-	scheme := runtime.NewScheme()
-	err := corev1.AddToScheme(scheme)
+	scheme, err := kube.CreateScheme()
 	if err != nil {
-		t.Fatal(err)
+		t.Fatalf("error creating fake client: %v", err)
 	}
 
 	return fake.NewClientBuilder().
 		WithScheme(scheme).
 		WithRuntimeObjects(clusterState...).
 		Build()
+}
+
+// CreateLogger create logger
+func CreateLogger() logger.Logger {
+	return logger.NewCLILogger(os.Stdout)
 }
